fix get accounts

+++ models/account.go
@@ -8,7 +8,7 @@ type Account struct {
 	ID           uint          `json:"-"`
 	UUID         uuid.UUID     `json:"uuid"`
 	IBAN         string        `json:"iban"`
-	Balance      float64       `json:"balance"`
+	Balance      uint          `json:"balance"`
 	UserId       uint          `json:"user_id"`
 	Sources      []Transaction `json:"sources"`
 	Destinations []Transaction `json:"destination"`
